+++ packages/awino-ui/src/components/pages/dashboard/InfoSection/InfoSection.test.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+
+import nextUseRouterMock from '@/mocks/nextUseRouterMock';
+import { cleanup, render } from '@/testing/utils';
+
+import InfoSection from './InfoSection';
+
+beforeAll(() => {
+  nextUseRouterMock({
+    route: '/',
+    pathname: '/',
+    query: '',
+    asPath: '',
+  });
+});
+
+afterEach(cleanup);
+
+describe('<InfoSection />', () => {
+  it('has valid snapshot', () => {
+    const { asFragment } = render(<InfoSection />, {}, 'dashboard');
+    expect(asFragment()).toMatchSnapshot();
+  });
+});